+++ client/src/signal.ts
@@ -1,0 +1,7 @@
+import { Assessment } from "./assessment";
+
+export type SignalPayload = 
+| {
+    type: "NewAssessment",
+    assessment: Assessment,
+}