+++ dnas/sensemaker/zomes/coordinator/sensemaker/src/lib.rs
@@ -8,6 +8,8 @@ mod range;
 mod resource_def;
 mod sensemaker_config;
 mod utils;
+mod signals;
+mod agent;
 pub use applet::*;
 pub use assessment::*;
 pub use cultural_context::*;