+++ dnas/sensemaker/zomes/integrity/sensemaker/src/lib.rs
@@ -45,6 +45,7 @@ pub enum LinkTypes {
     AppletName,
     AppletConfig,
     ResourceDefEhToAppletConfig,
+    AllAgentsPath,
 }
 
 #[hdk_extern]