+++ client/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 All notable changes to the `@neighbourhoods/client` package will be documented in this file.
 
+## v0.0.2 - 2023-05-09
+- `SensemakerStore` constructor inputs changed from `Sensemaker(public service: SensemakerService)` to `constructor(public client: AppAgentClient, public roleName: RoleName, public zomeName = 'sensemaker')` so that the signal handler can be set up in the constructor.
+- signal handler implemented so that whenever a signal of type `NewAssessment` is received, the assessment store is updated.
 ## v0.0.1 - 2023-04-18
 - Initial release under new name
 - Types and api methods updated to reflect [these changes](../CHANGELOG.md#v006-alpha---2023-04-18)